# services/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/website/gallery": {
      "get": {
        "operationId": "GalleryController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "gallery"
        ]
      },
      "post": {
        "operationId": "GalleryController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGalleryItemDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "gallery"
        ]
      }
    },
    "/website/gallery/{id}": {
      "get": {
        "operationId": "GalleryController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "gallery"
        ]
      },
      "patch": {
        "operationId": "GalleryController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGalleryItemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "gallery"
        ]
      },
      "delete": {
        "operationId": "GalleryController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "gallery"
        ]
      }
    },
...
  },
  "info": {
    "title": "Aram API",
    "description": "The Aram API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateGalleryItemDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "altText": {
            "type": "string"
          },
          "imagePath": {
            "type": "string"
          },
          "albumId": {
            "type": "number"
          },
          "tagsJson": {
            "type": "string"
          },
          "sortOrder": {
            "type": "number"
          },
          "caption": {
            "type": "string"
          },
          "photographer": {
            "type": "string"
          },
          "visibility": {
            "type": "string"
          },
          "fileSize": {
            "type": "string"
          },
          "thumbnailPath": {
            "type": "string"
          },
          "uploadedBy": {
            "type": "string"
          },
          "altTextGeneratedByAi": {
            "type": "boolean"
          }
        },
        "required": [
          "title",
          "imagePath"
        ]
      },
      "UpdateGalleryItemDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "altText": {
            "type": "string"
          },
          "imagePath": {
            "type": "string"
          },
          "albumId": {
            "type": "number"
          },
          "tagsJson": {
            "type": "string"
          },
          "sortOrder": {
            "type": "number"
          },
          "caption": {
            "type": "string"
          },
          "photographer": {
            "type": "string"
          },
          "visibility": {
            "type": "string"
          },
          "fileSize": {
            "type": "string"
          },
          "thumbnailPath": {
            "type": "string"
          },
          "uploadedBy": {
            "type": "string"
          },
          "altTextGeneratedByAi": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}